# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/evaluate/project/{pid}/training/{tid}/evaluate": {
            "post": {
                "description": "create evaluate",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Evaluate"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "training id",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body of creating inference",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.EvaluateCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.EvaluateDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bad_request_body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "bad_request_param"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/evaluate/project/{pid}/training/{tid}/evaluate/{id}": {
            "get": {
                "description": "watch evaluate",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Evaluate"
                ],
                "summary": "Watch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "training id",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "evaluate id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.EvaluateDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bad_request_body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "bad_request_param"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.EvaluateDTO": {
            "type": "object",
            "properties": {
                "access_url": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "evaluate_id": {
                    "type": "string"
                }
            }
        },
...
        "controller.EvaluateCreateRequest": {
            "type": "object",
            "properties": {
                "batch_size_scope": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "learning_rate_scope": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "momentum_scope": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}